Completed E-Commerce Backend MVP
Implemented Authentication, Brand, Product, Product Variant,
Product Images, Cart, Wishlist, Orders, Reviews, Search,
Filtering, Sorting, Pagination and Product Details APIs.
Integrated JWT Security, Hibernate ORM, DTO-Service-DAO architecture,
stock management, rating calculations and ownership validations.

+++ src/main/java/entity/Product.java
@@ -125,7 +125,7 @@ public class Product {
 		this.averageRating = averageRating;
 	}
 
-	public double getReviewCount() {
+	public int getReviewCount() {
 		return reviewCount;
 	}
